include wheel moves extraction in the pipeline
