+++ ibllib/io/extractors/biased_trials.py
@@ -5,7 +5,7 @@ import ibllib.io.raw_data_loaders as raw
 from ibllib.io.extractors.training_trials import (  # noqa; noqa
     CameraTimestamps, Choice, FeedbackTimes, FeedbackType, GoCueTimes, GoCueTriggerTimes,
     IncludedTrials, Intervals, ItiDuration, ProbabilityLeft, ResponseTimes, RewardVolume,
-    StimOnTimes, StimOnTriggerTimes, FirstMovementTimes)
+    StimOnTimes, StimOnTriggerTimes, StimOnOffFreezeTimes)
 from ibllib.misc import version
 
 
@@ -33,7 +33,7 @@ def extract_all(session_path, save=False, bpod_trials=False, settings=False):
         settings = {'IBLRIG_VERSION_TAG': '100.0.0'}
     base = [FeedbackType, ContrastLR, ProbabilityLeft, Choice, RewardVolume,
             FeedbackTimes, StimOnTimes, Intervals, ResponseTimes, GoCueTriggerTimes,
-            GoCueTimes, FirstMovementTimes, CameraTimestamps]
+            GoCueTimes, CameraTimestamps]
     # Version specific extractions
     if version.ge(settings['IBLRIG_VERSION_TAG'], '5.0.0'):
         base.extend([StimOnTriggerTimes, IncludedTrials])
