+++ ibllib/io/extractors/ephys_fpga.py
@@ -18,7 +18,6 @@ import ibllib.dsp as dsp
 from ibllib.io.extractors.base import BaseBpodTrialsExtractor, BaseExtractor, run_extractor_classes
 from ibllib.io.extractors import biased_trials
 from ibllib.io.extractors.training_wheel import extract_wheel_moves
-from ibllib.io.extractors.training_trials import FirstMovementTimes
 
 
 _logger = logging.getLogger('ibllib')
@@ -551,18 +550,6 @@ class ProbaContrasts(BaseBpodTrialsExtractor):
         return pe['prob_left'], pe['contrast_left'], pe['contrast_right']
 
 
-class Wheel(BaseExtractor):
-    save_names = ('_ibl_wheel.timestamps.npy', '_ibl_wheel.position.npy',
-                  '_ibl_wheelMoves.intervals.npy', '_ibl_wheelMoves.peakAmplitude.npy')
-    var_names = ('wheel_timestamps', 'wheel_position',
-                 'wheelMoves_intervals', 'wheelMoves_peakAmplitude')
-
-    def _extract(self, sync=None, chmap=None):
-        ts, pos = extract_wheel_sync(sync=sync, chmap=chmap)
-        moves = extract_wheel_moves(ts, pos)
-        return ts, pos, moves['intervals'], moves['peakAmplitude']
-
-
 class CameraTimestamps(BaseExtractor):
     save_names = ['_ibl_rightCamera.times.npy', '_ibl_leftCamera.times.npy',
                   '_ibl_bodyCamera.times.npy']
@@ -581,13 +568,17 @@ class FpgaTrials(BaseExtractor):
                   '_ibl_trials.response_times.npy', '_ibl_trials.goCueTrigger_times.npy',
                   '_ibl_trials.stimOn_times.npy', '_ibl_trials.stimOff_times.npy',
                   '_ibl_trials.goCue_times.npy', '_ibl_trials.feedback_times.npy',
-                  '_ibl_trials.firstMovement_times.npy')
+                  '_ibl_trials.firstMovement_times.npy', '_ibl_wheel.timestamps.npy',
+                  '_ibl_wheel.position.npy', '_ibl_wheelMoves.intervals.npy',
+                  '_ibl_wheelMoves.peakAmplitude.npy')
     var_names = ('probabilityLeft', 'contrastLeft', 'contrastRight', 'feedbackType', 'choice',
                  'rewardVolume', 'intervals_bpod', 'intervals', 'response_times',
                  'goCueTrigger_times', 'stimOn_times', 'stimOff_times', 'goCue_times',
-                 'feedback_times', 'firstMovement_times')
+                 'feedback_times', 'firstMovement_times', 'wheel_timestamps', 'wheel_position',
+                 'wheelMoves_intervals', 'wheelMoves_peakAmplitude')
 
     def _extract(self, sync=None, chmap=None):
+        # extracts trials
         # extract the behaviour data from bpod
         if sync is None or chmap is None:
             _sync, _chmap = _get_main_probe_sync(self.session_path, bin_exists=False)
@@ -610,29 +601,27 @@ class FpgaTrials(BaseExtractor):
         fpga_fields = ['stimOn_times', 'stimOff_times', 'goCue_times', 'feedback_times']
         # get ('probabilityLeft', 'contrastLeft', 'contrastRight') from the custom ephys extractors
         pclcr, _ = ProbaContrasts(self.session_path).extract(bpod_trials=bpod_raw, save=False)
-        # Get first movement times using wheel moves data
-        first_move_onsets = None
-        try:
-            wheel_moves = alf.io.load_object(self.session_path / 'alf', '_ibl_wheelMoves')
-            first_move_onsets, *_ = FirstMovementTimes(self.session_path)\
-                .extract(wheel_moves=wheel_moves, trials=fpga_trials, save=False)
-        except FileNotFoundError:
-            _logger.error('failed to load wheelMoves object for firstMovement extraction')
-        except Exception:
-            _logger.exception('failed to extract firstMovement times')
-
+        # build trials output
         out = OrderedDict()
-        out.update({k: pclcr[i] for i, k in enumerate(ProbaContrasts.var_names)})
+        out.update({k: pclcr[i][ifpga] for i, k in enumerate(ProbaContrasts.var_names)})
         out.update({k: bpod_trials[k][ibpod] for k in bpod_fields})
         out.update({k: fcn_bpod2fpga(bpod_trials[k][ibpod]) for k in bpod_rsync_fields})
         out.update({k: fpga_trials[k][ifpga] for k in fpga_fields})
+
+        # extract the wheel data
+        from ibllib.io.extractors.training_wheel import extract_first_movement_times
+        ts, pos = extract_wheel_sync(sync=sync, chmap=chmap)
+        moves = extract_wheel_moves(ts, pos)
+        settings = raw_data_loaders.load_settings(session_path=self.session_path)
+        min_qt = settings.get('QUIESCENT_PERIOD', None)
+        first_move_onsets, *_ = extract_first_movement_times(moves, out, min_qt=min_qt)
         out.update({'firstMovement_times': first_move_onsets})
 
-        assert self.var_names == tuple(out.keys())
-        return [out[k] for k in out]
+        assert tuple(filter(lambda x: 'wheel' not in x, self.var_names)) == tuple(out.keys())
+        return [out[k] for k in out] + [ts, pos, moves['intervals'], moves['peakAmplitude']]
 
 
-def extract_all(session_path, save=False, bin_exists=False):
+def extract_all(session_path, save=True, bin_exists=False):
     """
     For the IBL ephys task, reads ephys binary file and extract:
         -   sync
@@ -644,8 +633,9 @@ def extract_all(session_path, save=False, bin_exists=False):
     :param version: bpod version, defaults to None
     :return: outputs, files
     """
+    assert save  # fixme with wheel positions, this function can't work without saving the data
     sync, chmap = _get_main_probe_sync(session_path, bin_exists=bin_exists)
     outputs, files = run_extractor_classes(
-        [Wheel, CameraTimestamps, FpgaTrials], session_path=session_path,
+        [CameraTimestamps, FpgaTrials], session_path=session_path,
         save=save, sync=sync, chmap=chmap)
     return outputs, files
