+++ ibllib/io/extractors/training_wheel.py
@@ -1,8 +1,10 @@
 import logging
+from collections import Sized
 
 import numpy as np
 from scipy import interpolate
 
+from ibllib.io.extractors import training_trials
 from ibllib.io.extractors.base import BaseBpodTrialsExtractor
 import ibllib.io.raw_data_loaders as raw
 from ibllib.misc import structarr
@@ -252,7 +254,7 @@ def extract_wheel_moves(re_ts, re_pos, display=False):
     # [rad_X4, rad_X2, rad_X1, cm_X4, cm_X2, cm_X1]
     min_change = np.concatenate([2 * np.pi / res, wh.WHEEL_DIAMETER * np.pi / res])
     pos_diff = np.abs(np.ediff1d(re_pos)).min()
-    
+
     # find min change closest to min pos_diff
     idx = np.argmin(np.abs(min_change - pos_diff))
     if idx < len(res):
@@ -290,6 +292,53 @@ def extract_wheel_moves(re_ts, re_pos, display=False):
     return wheel_moves
 
 
+def extract_first_movement_times(wheel_moves, trials, min_qt=None):
+    """
+    Extracts the time of the first sufficiently large wheel movement for each trial.
+    To be counted, the movement must occur between go cue / stim on and before feedback /
+    response time.  The movement onset is sometimes just before the cue (occurring in the
+    gap between quiescence end and cue start, or during the quiescence period but sub-
+    threshold).  The movement is sufficiently large if it is greater than or equal to THRESH
+    :param wheel_moves: dictionary of detected wheel movement onsets and peak amplitudes for
+    use in extracting each trial's time of first movement.
+    :param trials: dictionary of trial data
+    :param min_qt: the minimum quiescence period, if None a default is used
+    :return: numpy array of first movement times, bool array indicating whether movement
+    crossed response threshold, and array of indices for wheel_moves arrays
+    """
+    THRESH = .1  # peak amp should be at least .1 rad; ~1/3rd of the distance to threshold
+    MIN_QT = .2  # default minimum enforced quiescence period
+
+    # Determine minimum quiescent period
+    if min_qt is None:
+        min_qt = MIN_QT
+        _logger.info('minimum quiescent period assumed to be %.0fms', MIN_QT * 1e3)
+    elif isinstance(min_qt, Sized) and len(min_qt) > len(trials['goCue_times']):
+        min_qt = np.array(min_qt[0:trials['goCue_times'].size])
+
+    # Initialize as nans
+    first_move_onsets = np.full(trials['goCue_times'].shape, np.nan)
+    ids = np.full(trials['goCue_times'].shape, int(-1))
+    is_final_movement = np.zeros(trials['goCue_times'].shape, bool)
+    flinch = abs(wheel_moves['peakAmplitude']) < THRESH
+    all_move_onsets = wheel_moves['intervals'][:, 0]
+    # Iterate over trials, extracting onsets approx. within closed-loop period
+    for i, (t1, t2) in enumerate(zip(trials['goCue_times'] - min_qt,
+                                     trials['feedback_times'])):
+        if ~np.isnan(t2 - t1):  # If both timestamps defined
+            mask = (all_move_onsets > t1) & (all_move_onsets < t2)
+            if np.any(mask):  # If any onsets for this trial
+                trial_onset_ids, = np.where(mask)
+                if np.any(~flinch[mask]):  # If any trial moves were sufficiently large
+                    ids[i] = trial_onset_ids[~flinch[mask]][0]  # Find first large move id
+                    first_move_onsets[i] = all_move_onsets[ids[i]]  # Save first large onset
+                    is_final_movement[i] = ids[i] == trial_onset_ids[-1]  # Final move of trial
+        else:  # Log missing timestamps
+            _logger.warning('no reliable times for trial id %i', i + 1)
+
+    return first_move_onsets, is_final_movement, ids[ids != -1]
+
+
 class Wheel(BaseBpodTrialsExtractor):
     """
     Get wheel data from raw files and converts positions into radians mathematical convention
@@ -304,14 +353,26 @@ class Wheel(BaseBpodTrialsExtractor):
     Radians mathematical convention
     """
     save_names = ('_ibl_wheel.timestamps.npy', '_ibl_wheel.position.npy',
-                  '_ibl_wheelMoves.intervals.npy', '_ibl_wheelMoves.peakAmplitude.npy')
-    var_names = ('wheel_timestamps', 'wheel_position',
-                 'wheel_moves_intervals', 'wheel_moves_peak_amplitude')
+                  '_ibl_wheelMoves.intervals.npy', '_ibl_wheelMoves.peakAmplitude.npy',
+                  '_ibl_trials.firstMovement_times.npy')
+    var_names = ('wheel_timestamps', 'wheel_position', 'wheel_moves_intervals',
+                 'wheel_moves_peak_amplitude', 'firstMovement_times')
 
     def _extract(self):
         ts, pos = get_wheel_position(self.session_path, self.bpod_trials)
         moves = extract_wheel_moves(ts, pos)
-        return ts, pos, moves['intervals'], moves['peakAmplitude']
+
+        # need some trial based info to output the first movement times
+        goCue_times, _ = training_trials.GoCueTimes(self.session_path).extract(
+            save=False, bpod_trials=self.bpod_trials, settings=self.settings)
+        feedback_times, _ = training_trials.FeedbackTimes(self.session_path).extract(
+            save=False, bpod_trials=self.bpod_trials, settings=self.settings)
+        trials = {'goCue_times': goCue_times, 'feedback_times': feedback_times}
+        min_qt = self.settings.get('QUIESCENT_PERIOD', None)
+
+        first_moves, *_ = extract_first_movement_times(moves, trials, min_qt=min_qt)
+
+        return ts, pos, moves['intervals'], moves['peakAmplitude'], first_moves
 
 
 def extract_all(session_path, bpod_trials=None, settings=None, save=False):
